% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inter_module.R
\name{abundanceHeatmap}
\alias{abundanceHeatmap}
\title{Abundance Heatmap module - server}
\usage{
abundanceHeatmap(
  input,
  output,
  session,
  aggDat,
  featLevel,
  colorOptions,
  levelOpts,
  hmSort,
  hmFeatList,
  reset
)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{aggDat}{aggregated MRExperiment}

\item{featLevel}{chosen feature level (aggregation level)}

\item{colorOptions}{reactive storing filters selected via data input}

\item{levelOpts}{all available level choices for this dataset}

\item{hmSort}{reactive storing sorting method for heatmap}

\item{hmFeatList}{reactive storing list of features to include in heatmap}

\item{reset}{boolean reactive which resets the module if TRUE}
}
\value{
R code needed to generate the heatmap
}
\description{
Abundance Heatmap module - server
}
\author{
Janina Reeder
}
