% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inter_module.R
\name{abundanceHeatmapUI}
\alias{abundanceHeatmapUI}
\title{Abundance Heatmap module - UI}
\usage{
abundanceHeatmapUI(id)
}
\arguments{
\item{id}{namespace identifier}
}
\value{
box holding the UI code
}
\description{
Abundance Heatmap module - UI
}
\author{
Janina Reeder
}
