% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{add_plotly_config}
\alias{add_plotly_config}
\title{Adds a config call based on plotly::config}
\usage{
add_plotly_config(.data)
}
\arguments{
\item{.data}{plotly data object to apply the config call to}
}
\value{
plotly::config call
}
\description{
Adds a config call based on plotly::config
}
