% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{add_plotly_layout}
\alias{add_plotly_layout}
\title{Adds a layout call based on plotly::layout}
\usage{
add_plotly_layout(.data, plotTitle, xaxis_text, ylab)
}
\arguments{
\item{.data}{plotly data object to apply the layout call to}

\item{plotTitle}{plot title to use}

\item{xaxis_text}{x axis label to use}

\item{ylab}{y axis label to use}
}
\value{
plotly::layout call
}
\description{
Adds a layout call based on plotly::layout
}
