% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataManipulationFunctions.R
\name{addData}
\alias{addData}
\title{Add Data to MicrobiomeData Object}
\usage{
addData(object, new_data, data_type)
}
\arguments{
\item{object}{MicrobiomeData object to add data to.}

\item{new_data}{New data to add (matrix or data frame).}

\item{data_type}{Type of the new data ('rRNA16S', 'Metagenomic', or 'Metatranscriptomic').}
}
\value{
Updated MicrobiomeData object.
}
\description{
Merges new data into an existing MicrobiomeData object with advanced handling
of overlapping samples and consistency checks.
}
