% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{addFeatData}
\alias{addFeatData}
\title{Add feature data to MRobj.}
\usage{
addFeatData(MRobj, featdata = NULL)
}
\arguments{
\item{MRobj}{An MRexperiment object.}

\item{featdata}{Feature data frame or file path.}
}
\value{
An updated MRexperiment object.
}
\description{
This function adds feature data to the featureData slot in an MRexperiment
object.
}
