% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{addPhenoData}
\alias{addPhenoData}
\title{Add phenotype data to object.}
\usage{
addPhenoData(MRobj, phenodata = NULL)
}
\arguments{
\item{MRobj}{An MRexperiment object.}

\item{phenodata}{Phenotype data frame or file path.}
}
\value{
An updated MRexperiment object.
}
\description{
This function adds phenotype data to the phenoData slot in an MRexperiment
object.
}
