% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation_module.R
\name{aggregationTab}
\alias{aggregationTab}
\title{Aggregation module server function}
\usage{
aggregationTab(
  input,
  output,
  session,
  resetInput,
  levelOpts,
  chosenLevel,
  meData
)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{resetInput}{boolean updated to TRUE if new data is available}

\item{levelOpts}{available levels to aggregate on (depends on input data)}

\item{chosenLevel}{previously selected level (passed from different instance)}

\item{meData}{the main MRexperiment object}
}
\value{
reactive list holding aggregated object, aggregation code and 
boolean on normalization
}
\description{
Aggregation module server function
}
\author{
Janina Reeder
}
