% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation_module.R
\name{aggregationTabUI}
\alias{aggregationTabUI}
\title{Aggregation module ui function}
\usage{
aggregationTabUI(id)
}
\arguments{
\item{id}{namespace identifier}
}
\value{
box holding aggregation input elements
}
\description{
Aggregation module ui function
}
\examples{
aggregationTabUI("atu_id")

}
\author{
Janina Reeder
}
