% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intra_module.R
\name{alphaDiversity}
\alias{alphaDiversity}
\title{Alpha Diversity module - server}
\usage{
alphaDiversity(
  input,
  output,
  session,
  aggDat,
  featLevel,
  intraSettings,
  colorOptions,
  reset
)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{aggDat}{aggregated MRExperiment}

\item{featLevel}{chosen feature level (aggregation level)}

\item{intraSettings}{analysis settings as passed over from analysis 
input module}

\item{colorOptions}{phenotype selections: used for color choices}

\item{reset}{boolean reactive which resets the module if TRUE}
}
\value{
R code used to make the alpha diversity plot
}
\description{
Alpha Diversity module - server
}
\author{
Janina Reeder
}
