% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intra_module.R
\name{alphaDiversityUI}
\alias{alphaDiversityUI}
\title{Alpha Diversity module - UI}
\usage{
alphaDiversityUI(id)
}
\arguments{
\item{id}{namespace identifier}
}
\value{
box holding the UI code
}
\description{
Alpha Diversity module - UI
}
\author{
Janina Reeder
}
