% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_main_module.R
\name{avgAbundanceUI}
\alias{avgAbundanceUI}
\title{Relative abundance plot module - UI}
\usage{
avgAbundanceUI(id)
}
\arguments{
\item{id}{namespace identifier}
}
\value{
box containing the ui code
}
\description{
Relative abundance plot module - UI
}
\author{
Janina Reeder
}
