% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inter_module.R
\name{betaDiversity}
\alias{betaDiversity}
\title{Beta Diversity module - server}
\usage{
betaDiversity(
  input,
  output,
  session,
  aggDat,
  aggLevel,
  colorOptions,
  shapeOptions,
  betadistance,
  betaSettings,
  reset
)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{aggDat}{MRExperiment storing data}

\item{aggLevel}{aggregation level}

\item{colorOptions}{phenotype selection options for color}

\item{shapeOptions}{phenotype selection options for shape}

\item{betadistance}{distance measured used for beta diversity analysis}

\item{betaSettings}{input choices for beta diversity}

\item{reset}{boolean reactive which resets the module if TRUE}
}
\value{
R code needed to generate the beta diversity plot
}
\description{
Beta Diversity module - server
}
\author{
Janina Reeder
}
