% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inter_module.R
\name{betaDiversityUI}
\alias{betaDiversityUI}
\title{Beta Diversity module - UI}
\usage{
betaDiversityUI(id)
}
\arguments{
\item{id}{namespace identifier}
}
\value{
box holding the ui code
}
\description{
Beta Diversity module - UI
}
\author{
Janina Reeder
}
