% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inter_input_module.R
\name{betaInput}
\alias{betaInput}
\title{Server side for the analysis input module handling analysis control}
\usage{
betaInput(input, output, session, meData, adonisOptions, reset)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{meData}{MRExperiment object storing all data}

\item{adonisOptions}{phenodata colums ready for adonis analysis}

\item{reset}{reactive boolean determining if all inputs should be reset}
}
\value{
list holding all chosen values and the selected feature
}
\description{
Server side for the analysis input module handling analysis control
}
\author{
Janina Reeder
}
