% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inter_input_module.R
\name{betaInputUI}
\alias{betaInputUI}
\title{Main beta analysis input module.
Set up to handle all analysis tabs in the app depending on given parameters}
\usage{
betaInputUI(id)
}
\arguments{
\item{id}{element identifier - namespace}
}
\value{
box containing ui element
}
\description{
Main beta analysis input module.
Set up to handle all analysis tabs in the app depending on given parameters
}
\author{
Janina Reeder
}
