% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{buildEmptyPlotlyPlot}
\alias{buildEmptyPlotlyPlot}
\title{Creates an empty plotly plot using the given labels on the x and y axis}
\usage{
buildEmptyPlotlyPlot(xaxis_text, ylab)
}
\arguments{
\item{xaxis_text}{x axis label}

\item{ylab}{y axis label}
}
\value{
call to plotly_empty
}
\description{
Creates an empty plotly plot using the given labels on the x and y axis
}
