% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_functions.R
\name{computeCI_Interval}
\alias{computeCI_Interval}
\title{Helper function to calculate the confidence interval for a cor.test}
\usage{
computeCI_Interval(num, mS, method)
}
\arguments{
\item{num}{number of samples}

\item{mS}{results of cor.test}

\item{method}{statistical method used for cor.test}
}
\value{
named vector holding lower and upper thresholds
}
\description{
Helper function to calculate the confidence interval for a cor.test
}
