% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_analysis_module.R
\name{corrAnalysisUI}
\alias{corrAnalysisUI}
\title{corr Analysis Module - UI}
\usage{
corrAnalysisUI(id)
}
\arguments{
\item{id}{namespace identifier}
}
\value{
fluidRow containing the ui code
}
\description{
corr Analysis Module - UI
}
\examples{
corrAnalysisUI("coranalysis_id")

}
\author{
Janina Reeder
}
