% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_input_module.R
\name{corrInputUI}
\alias{corrInputUI}
\title{Main correlation analysis input module.
Handles correlation analysis tab in the app}
\usage{
corrInputUI(id, type)
}
\arguments{
\item{id}{element identifier - namespace}

\item{type}{determines if 'feature' or 'pheno' correlation}
}
\value{
box containing ui element
}
\description{
Main correlation analysis input module.
Handles correlation analysis tab in the app
}
\author{
Janina Reeder
}
