% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{createHeader}
\alias{createHeader}
\title{Makes header for R script}
\usage{
createHeader(
  title = "MicrobiomeExplorer Report",
  author = "",
  date = "",
  data.source = "",
  output = getOption("me.reportformat"),
  toc = TRUE
)
}
\arguments{
\item{title}{Title of the report.}

\item{author}{Author of the report.}

\item{date}{Date of the report.}

\item{data.source}{R code used to obtain the dataset}

\item{output}{Output of Rmarkdown file.}

\item{toc}{Table of contents. Default is TRUE.}
}
\value{
A character vector where each element is a line in the R script.
}
\description{
This function makes the header for the report R script to be rendered by
knitr into Rmarkdown and rendered into HTML, PDF, or Word.
}
\details{
This was adapted from https://yihui.name/knitr/demo/stitch/
}
