% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiomeDataClass.R
\name{createMicrobiomeDataObject}
\alias{createMicrobiomeDataObject}
\alias{createMicrobiomeDataObject,ANY,ANY,ANY,list-method}
\title{Create a new MicrobiomeData object}
\usage{
createMicrobiomeDataObject(
  rRNA16S,
  Metagenomic,
  Metatranscriptomic,
  SampleInfo,
  PhylogeneticTree = NULL
)

\S4method{createMicrobiomeDataObject}{ANY,ANY,ANY,list,ANY}(
  rRNA16S,
  Metagenomic,
  Metatranscriptomic,
  SampleInfo,
  PhylogeneticTree = NULL
)
}
\arguments{
\item{rRNA16S}{A matrix or dataframe with 16S rRNA gene sequencing data.}

\item{Metagenomic}{A matrix or dataframe with metagenomic data.}

\item{Metatranscriptomic}{A matrix or dataframe with metatranscriptomic data.}

\item{SampleInfo}{A data frame with sample information.}

\item{PhylogeneticTree}{A phylogenetic tree object used by tree-aware analyses.}
}
\value{
A `MicrobiomeData` object.
}
\description{
This function initializes a MicrobiomeData object with the provided data.
}
\examples{
# Example usage of createMicrobiomeDataObject

# Creating mock data for the example
myRNA16SData <- matrix(runif(50, min = 0, max = 10), nrow = 5, ncol = 10)
colnames(myRNA16SData) <- paste0("OTU", 1:10)
rownames(myRNA16SData) <- paste0("Sample", 1:5)

myMetagenomicData <- matrix(rnorm(20, mean = 10), nrow = 5, ncol = 4)
colnames(myMetagenomicData) <- paste0("Gene", 1:4)

myMetatranscriptomicData <- matrix(rnorm(20, mean = 5), nrow = 5, ncol = 4)
colnames(myMetatranscriptomicData) <- paste0("Expression", 1:4)

mySampleInfo <- data.frame(
  SampleID = paste0("Sample", 1:5),
  Condition = rep(c("Control", "Treatment"), length.out = 5)
)
rownames(mySampleInfo) <- mySampleInfo$SampleID

# Using the function with the mock data
myData <- createMicrobiomeDataObject(
  rRNA16S = myRNA16SData,
  Metagenomic = myMetagenomicData,
  Metatranscriptomic = myMetatranscriptomicData,
  SampleInfo = mySampleInfo
)

}
