% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_input_module.R
\name{dataInput}
\alias{dataInput}
\title{Main Data input server where the user selects files to upload to the app or 
connects to database}
\usage{
dataInput(
  input,
  output,
  session,
  dataSource,
  dataFilterRep,
  qcRep,
  addPheno,
  resetReports
)
}
\arguments{
\item{input}{module input}

\item{output}{module output}

\item{session}{app session}

\item{dataSource}{reactive Value storing commands for loading data}

\item{dataFilterRep}{reactive Value storing commands for filtering data}

\item{qcRep}{reactive Value storing commands for producing qc plots}

\item{addPheno}{reactive boolean keeping track of phenodata changes}

\item{resetReports}{reactive boolean indicating whether reports need to be reset}
}
\value{
list of reactives containing the uploaded and filtered data as well 
as the filterChoices on phenotypes
}
\description{
Main Data input server where the user selects files to upload to the app or 
connects to database
}
\author{
Janina Reeder
}
