% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_input_module.R
\name{dataInputUI}
\alias{dataInputUI}
\title{Main Data input UI where the user selects files to upload to the app or 
connects to database}
\usage{
dataInputUI(id)
}
\arguments{
\item{id}{module identifier}
}
\value{
fluidRow holding UI interface
}
\description{
Main Data input UI where the user selects files to upload to the app or 
connects to database
}
\examples{
dataInputUI("datainput_id")

}
\author{
Janina Reeder
}
