% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_functions.R
\name{designPairs}
\alias{designPairs}
\title{Produce design matrix of pairwise comparisons}
\usage{
designPairs(levels)
}
\arguments{
\item{levels}{Character vector of the levels of a factor phenotype}
}
\value{
A model matrix
}
\description{
This function takes in the levels of a factor phenotype and outputs a
design matrix of all pairwise comparisons.
}
