% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_analysis_module.R
\name{diffAnalysisUI}
\alias{diffAnalysisUI}
\title{Diff Analysis Module - UI}
\usage{
diffAnalysisUI(id)
}
\arguments{
\item{id}{namespace identifier}
}
\value{
fluidRow containing the ui code
}
\description{
Diff Analysis Module - UI
}
\examples{
diffAnalysisUI("diffanalysis_id")

}
\author{
Janina Reeder
}
