% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_input_module.R
\name{diffInput}
\alias{diffInput}
\title{Server side for the analysis input module handling analysis control}
\usage{
diffInput(input, output, session, meData, facetOptions = NULL, reset)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{meData}{MRExperiment object storing all data}

\item{facetOptions}{named vector of available facet choices}

\item{reset}{reactive boolean determining if all inputs should be reset}
}
\value{
list holding all chosen values and the selected feature
}
\description{
Server side for the analysis input module handling analysis control
}
\author{
Janina Reeder
}
