% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_input_module.R
\name{diffInputUI}
\alias{diffInputUI}
\title{Main diffanalysis input module.
Set up to handle diff analysis tabs in the app depending on given parameters}
\usage{
diffInputUI(id)
}
\arguments{
\item{id}{element identifier - namespace}
}
\value{
box containing ui element
}
\description{
Main diffanalysis input module.
Set up to handle diff analysis tabs in the app depending on given parameters
}
\author{
Janina Reeder
}
