% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_module.R
\name{diffTable}
\alias{diffTable}
\title{Differential analysis module server code}
\usage{
diffTable(
  input,
  output,
  session,
  aggDat,
  featLevel,
  diffSettings,
  reset,
  normalized
)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{aggDat}{aggregated MRExperiment}

\item{featLevel}{chosen feature level (aggregation level)}

\item{diffSettings}{reactive storing values selected in analysis 
input interface}

\item{reset}{boolean reactive which resets the module if TRUE}

\item{normalized}{boolean reactive indicating if data has been normalized}
}
\value{
list containing R code for analysis and for feature plots
}
\description{
Differential analysis module server code
}
\author{
Janina Reeder
}
