% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_module.R
\name{diffTableUI}
\alias{diffTableUI}
\title{Differential Analysis module UI}
\usage{
diffTableUI(id)
}
\arguments{
\item{id}{namespace identifier}
}
\value{
row containing the UI elements
}
\description{
Differential Analysis module UI
}
\author{
Janina Reeder
}
