% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{extendPhenoData}
\alias{extendPhenoData}
\title{Extends existing phenodata for an object}
\usage{
extendPhenoData(MRobj, phenodata = NULL)
}
\arguments{
\item{MRobj}{An MRexperiment object.}

\item{phenodata}{Phenotype data frame or file path.}
}
\value{
An updated MRexperiment object.
}
\description{
This function adds phenotype data to the phenoData slot in an MRexperiment
object.
}
