% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intra_module.R
\name{featAbundance}
\alias{featAbundance}
\title{Feature plot module - server}
\usage{
featAbundance(
  input,
  output,
  session,
  aggDat,
  featLevel,
  intraSettings,
  selectedFeat,
  featName,
  numOfFeats,
  ylabMode,
  normalizedData,
  reset
)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{aggDat}{aggregated MRExperiment}

\item{featLevel}{chosen feature level (aggregation level)}

\item{intraSettings}{analysis settings passed over from analysis input module}

\item{selectedFeat}{feature selected via drop down element of analysis input}

\item{featName}{plotly click event passed via relative abundance}

\item{numOfFeats}{number of features shown in relative abundance plot 
(affects plotly click data)}

\item{ylabMode}{character indication if raw \"Reads\" or \"Percentage\" 
should be shown}

\item{normalizedData}{boolean indicating whether data has been normalized}

\item{reset}{boolean reactive which resets the module if TRUE}
}
\value{
R code needed to build the feature plot
}
\description{
Feature plot module - server
}
\author{
Janina Reeder
}
