% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intra_module.R
\name{featAbundanceUI}
\alias{featAbundanceUI}
\title{Feature plot module - UI}
\usage{
featAbundanceUI(id)
}
\arguments{
\item{id}{namespace identifier}
}
\value{
box holding the UI code
}
\description{
Feature plot module - UI
}
