% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_analysis_module.R
\name{featureAnalysis}
\alias{featureAnalysis}
\title{feature Analysis Module - server}
\usage{
featureAnalysis(
  input,
  output,
  session,
  data,
  resetInput,
  aggData,
  normalizedData
)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{data}{the main data object returned from data_input_module}

\item{resetInput}{reactive boolean determining if reset is required}

\item{aggData}{the aggregated MRExperiment object}

\item{normalizedData}{boolean indicating if normalization was done}
}
\value{
reactive holding code to be used in reports
}
\description{
feature Analysis Module - server
}
\author{
Janina Reeder
}
