% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_analysis_module.R
\name{featureAnalysisUI}
\alias{featureAnalysisUI}
\title{feature Analysis Module - UI}
\usage{
featureAnalysisUI(id)
}
\arguments{
\item{id}{namespace identifier}
}
\value{
fluidRow containing the ui code
}
\description{
feature Analysis Module - UI
}
\examples{
featureAnalysisUI("featureanalysis_id")

}
\author{
Janina Reeder
}
