% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_module.R
\name{featureCorr}
\alias{featureCorr}
\title{Feature correlation analysis server module}
\usage{
featureCorr(
  input,
  output,
  session,
  aggDat,
  colorOptions,
  corFeatBase,
  corFeat2,
  corFacet1,
  corFacet2,
  corMethod,
  reset
)
}
\arguments{
\item{input}{module input}

\item{output}{module output}

\item{session}{app session}

\item{aggDat}{aggregated MRExperiment}

\item{colorOptions}{reactive storing filters available via data input}

\item{corFeatBase}{first correlation feature}

\item{corFeat2}{second correlation feature}

\item{corFacet1}{first correlation facet}

\item{corFacet2}{second correlation facet}

\item{corMethod}{correlation method to use}

\item{reset}{boolean reactive which resets the module if TRUE}
}
\value{
R code used to do the correlation analysis (character)
}
\description{
Feature correlation analysis server module
}
\author{
Janina Reeder
}
