% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_module.R
\name{featureCorrUI}
\alias{featureCorrUI}
\title{Feature correlation analysis module UI}
\usage{
featureCorrUI(id)
}
\arguments{
\item{id}{namespace identifier}
}
\value{
box containing the UI elements
}
\description{
Feature correlation analysis module UI
}
\author{
Janina Reeder
}
