% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_input_module.R
\name{featureInputUI}
\alias{featureInputUI}
\title{Main feature analysis input module.
Set up to handle all analysis tabs in the app depending on given parameters}
\usage{
featureInputUI(id)
}
\arguments{
\item{id}{element identifier - namespace}
}
\value{
box containing ui element
}
\description{
Main feature analysis input module.
Set up to handle all analysis tabs in the app depending on given parameters
}
\author{
Janina Reeder
}
