% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_module.R
\name{featureTable}
\alias{featureTable}
\title{Feature table module server code}
\usage{
featureTable(input, output, session, meData, featureModRep)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{meData}{MRExperiment storing the data}

\item{featureModRep}{reactiveValue storing modifications performed on fData}
}
\description{
Feature table module server code
}
\author{
Janina Reeder
}
