% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_module.R
\name{featureTableUI}
\alias{featureTableUI}
\title{Feature table UI module}
\usage{
featureTableUI(id)
}
\arguments{
\item{id}{namespace identifier}
}
\value{
fluidRow containing the UI code for feature tables
}
\description{
Feature table UI module
}
\author{
Janina Reeder
}
