% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_input_module.R
\name{fileUploadUI}
\alias{fileUploadUI}
\title{Module handling file upload for the application: UI
In a deployed version this module should be replaced with database access}
\usage{
fileUploadUI(id)
}
\arguments{
\item{id}{module identifier}
}
\description{
Module handling file upload for the application: UI
In a deployed version this module should be replaced with database access
}
\author{
Janina Reeder
}
