% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_module.R
\name{getFeatModCode}
\alias{getFeatModCode}
\title{Helper function returning the fData modifications as strings for 
report generation}
\usage{
getFeatModCode(featureanno)
}
\arguments{
\item{featureanno}{type of feature annotation; values are "Mark unknown" or 
"Roll down"}
}
\value{
String containing R code performing the modification
}
\description{
Helper function returning the fData modifications as strings for 
report generation
}
