% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_module.R
\name{getFeatSplitCode}
\alias{getFeatSplitCode}
\title{Helper function returning the fData modifications as strings for 
report generation}
\usage{
getFeatSplitCode(splittaxonomy)
}
\arguments{
\item{splittaxonomy}{name of column to split on}
}
\value{
String containing R code performing the modification
}
\description{
Helper function returning the fData modifications as strings for 
report generation
}
