% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getFileType}
\alias{getFileType}
\title{Helper function assigning different file extensions to specific short texts 
identifying the types}
\usage{
getFileType(fileext)
}
\arguments{
\item{fileext}{the file extension found after '.'}
}
\value{
character string for the filetype
}
\description{
Helper function assigning different file extensions to specific short texts 
identifying the types
}
\author{
Janina Reeder
}
