% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getLegendLevel}
\alias{getLegendLevel}
\title{Function to find a non-empty facet in the last row. This will be the
one to be connected to the plot legend to avoid duplicates within}
\usage{
getLegendLevel(df2, facets, facet2s)
}
\arguments{
\item{df2}{plotting data frame}

\item{facets}{column facets}

\item{facet2s}{row facets}
}
\value{
the name of the column-based facet which can be used as legend
}
\description{
Function to find a non-empty facet in the last row. This will be the
one to be connected to the plot legend to avoid duplicates within
}
