% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype_module.R
\name{getPhenoChanges}
\alias{getPhenoChanges}
\title{Helper function returning the code used to modify the data types of the
pheno table}
\usage{
getPhenoChanges(phenotype, datatype)
}
\arguments{
\item{phenotype}{name of the phenotype column header}

\item{datatype}{variable type to assign to column}
}
\value{
String storing code to perform modification
}
\description{
Helper function returning the code used to modify the data types of the
pheno table
}
\author{
Janina Reeder
}
