% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype_module.R
\name{getPhenoModCode}
\alias{getPhenoModCode}
\title{Helper function returning the code used to modify the phenotable as a string}
\usage{
getPhenoModCode(name, pheno1, pheno2)
}
\arguments{
\item{name}{interaction name}

\item{pheno1}{first interaction phenotype}

\item{pheno2}{second interaction phenotype}
}
\value{
String storing code to perform modification
}
\description{
Helper function returning the code used to modify the phenotable as a string
}
\author{
Janina Reeder
}
