% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getWidths}
\alias{getWidths}
\title{Helper function to account for issues plotly has with very small widths
(these end up being 1 and cover the entire plotting area)}
\usage{
getWidths(df2, facets, x_var, drop = TRUE)
}
\arguments{
\item{df2}{dataframe storing plotting data}

\item{facets}{column facets}

\item{x_var}{x variable}

\item{drop}{passed on as .drop to dplyr::group_by}
}
\value{
widths for each facet
}
\description{
Helper function to account for issues plotly has with very small widths
(these end up being 1 and cover the entire plotting area)
}
