% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inter_input_module.R
\name{heatmapInput}
\alias{heatmapInput}
\title{Server side for the analysis input module handling analysis control}
\usage{
heatmapInput(input, output, session, meData, reset, aggDat = reactive(NULL))
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{meData}{MRExperiment object storing all data}

\item{reset}{reactive boolean determining if all inputs should be reset}

\item{aggDat}{aggregated MRExperiment object (default is NULL)}
}
\value{
list holding all chosen values and the selected feature
}
\description{
Server side for the analysis input module handling analysis control
}
\author{
Janina Reeder
}
