% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inter_input_module.R
\name{heatmapInputUI}
\alias{heatmapInputUI}
\title{Heatmap analysis input module.
Set up to handle all analysis tabs in the app depending on given parameters}
\usage{
heatmapInputUI(id)
}
\arguments{
\item{id}{element identifier - namespace}
}
\value{
box containing ui element
}
\description{
Heatmap analysis input module.
Set up to handle all analysis tabs in the app depending on given parameters
}
\author{
Janina Reeder
}
