% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inter_analysis_module.R
\name{interAnalysisUI}
\alias{interAnalysisUI}
\title{inter Analysis Module - UI}
\usage{
interAnalysisUI(id)
}
\arguments{
\item{id}{namespace identifier}
}
\value{
fluidRow containing the ui code
}
\description{
inter Analysis Module - UI
}
\author{
Janina Reeder
}
