% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intra_analysis_module.R
\name{intraAnalysis}
\alias{intraAnalysis}
\title{Intra Analysis Module - server}
\usage{
intraAnalysis(
  input,
  output,
  session,
  data,
  levelOpts,
  chosenLevel,
  resetInput,
  aggData,
  normalizedData
)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{data}{the main data object returned from data_input_module}

\item{levelOpts}{available levels to aggregate on (depends on input data)}

\item{chosenLevel}{previously selected level (passed from different instance)}

\item{resetInput}{reactive boolean determining if reset is required}

\item{aggData}{the aggregated MRExperiment object}

\item{normalizedData}{boolean indicating if normalization was done}
}
\value{
reactive holding code to be used in reports
}
\description{
Intra Analysis Module - server
}
\author{
Janina Reeder
}
