% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intra_analysis_module.R
\name{intraAnalysisUI}
\alias{intraAnalysisUI}
\title{Intra Analysis Module - UI}
\usage{
intraAnalysisUI(id)
}
\arguments{
\item{id}{namespace identifier}
}
\value{
fluidRow containing the ui code
}
\description{
Intra Analysis Module - UI
}
\author{
Janina Reeder
}
