% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long_analysis_module.R
\name{longAnalysisUI}
\alias{longAnalysisUI}
\title{Long Analysis Module - UI}
\usage{
longAnalysisUI(id)
}
\arguments{
\item{id}{namespace identifier}
}
\value{
fluidRow containing the ui code
}
\description{
Long Analysis Module - UI
}
\author{
Janina Reeder
}
