% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long_input_module.R
\name{longInputUI}
\alias{longInputUI}
\title{Main diffanalysis input module.
Set up to handle diff analysis tabs in the app depending on given parameters}
\usage{
longInputUI(id)
}
\arguments{
\item{id}{element identifier - namespace}
}
\value{
box containing ui element
}
\description{
Main diffanalysis input module.
Set up to handle diff analysis tabs in the app depending on given parameters
}
\author{
Janina Reeder
}
