% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long_module.R
\name{longResults}
\alias{longResults}
\title{Longitudinal analysis module server code}
\usage{
longResults(
  input,
  output,
  session,
  aggDat,
  featLevel,
  longSettings,
  normalizedData,
  reset
)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{aggDat}{aggregated MRExperiment}

\item{featLevel}{chosen feature level (aggregation level)}

\item{longSettings}{reactive storing values selected in analysis 
input interface}

\item{normalizedData}{reactive boolean indicating if data has been normalized}

\item{reset}{boolean reactive which resets the module if TRUE}
}
\value{
list containing R code for analysis and for feature plots
}
\description{
Longitudinal analysis module server code
}
\author{
Janina Reeder
}
