% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long_module.R
\name{longResultsUI}
\alias{longResultsUI}
\title{Longitudinal Analysis module UI}
\usage{
longResultsUI(id)
}
\arguments{
\item{id}{namespace identifier}
}
\value{
row containing the UI elements
}
\description{
Longitudinal Analysis module UI
}
\author{
Janina Reeder
}
