% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MicrobiomeExplorer-package.R
\docType{package}
\name{MicrobiomeExplorer-package}
\alias{MicrobiomeExplorer}
\alias{MicrobiomeExplorer-package}
\title{MicrobiomeExplorer: Tools for Microbiome Analysis and Visualization}
\description{
MicrobiomeExplorer provides an integrated platform for the comprehensive analysis of microbiome data. It includes functionalities for preprocessing, statistical analysis, machine learning classification, biomarker discovery, network analysis, and advanced visualization, catering specifically to 16S rRNA gene sequencing, metagenomics, and metatranscriptomics data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/itsSabbir/MicrobiomeExplorer}
  \item Report bugs at \url{https://github.com/itsSabbir/MicrobiomeExplorer/issues}
}

}
\author{
\strong{Maintainer}: Sabbir Hossain \email{hossain.sabbir17@gmail.com}

}
\keyword{internal}
