% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{parseInteractionName}
\alias{parseInteractionName}
\title{Helper function used to build a correct interactionName based on the chosen 
columns}
\usage{
parseInteractionName(interactionName)
}
\arguments{
\item{interactionName}{as chosen by user. This may not be good to store 
internally}
}
\value{
updated interactionName or warning/error string
}
\description{
Helper function used to build a correct interactionName based on the chosen 
columns
}
