% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_module.R
\name{phenotypeCorr}
\alias{phenotypeCorr}
\title{Phenotype correlation analysis server module}
\usage{
phenotypeCorr(
  input,
  output,
  session,
  aggDat,
  colorOptions,
  corFeatBase,
  corPheno,
  corFacet1,
  corFacet2,
  corMethod,
  reset
)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{aggDat}{aggregated MRExperiment}

\item{colorOptions}{reactive storing filters available via data input}

\item{corFeatBase}{first correlation feature}

\item{corPheno}{correlation phenotype}

\item{corFacet1}{first correlation facet}

\item{corFacet2}{second correlation facet}

\item{corMethod}{correlation method to use}

\item{reset}{boolean reactive which resets the module if TRUE}
}
\value{
R code used to do the correlation analysis (character)
}
\description{
Phenotype correlation analysis server module
}
\author{
Janina Reeder
}
