% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_module.R
\name{phenotypeCorrUI}
\alias{phenotypeCorrUI}
\title{Phenotype correlation analysis module}
\usage{
phenotypeCorrUI(id)
}
\arguments{
\item{id}{namespace identifier}
}
\value{
box containing the UI element
}
\description{
Phenotype correlation analysis module
}
\author{
Janina Reeder
}
