% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype_module.R
\name{phenotypeTable}
\alias{phenotypeTable}
\title{Phenotype table server module}
\usage{
phenotypeTable(input, output, session, meData, phenoModRep, addPheno)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{meData}{MRExperiment storing the data}

\item{phenoModRep}{reactive Value storing any phenotable modifications made}

\item{addPheno}{reactive boolean keeping track of pheno data modifications}
}
\value{
phenotype table server fragment - no return value
}
\description{
Phenotype table server module
}
\author{
Janina Reeder
}
