% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype_module.R
\name{phenotypeTableUI}
\alias{phenotypeTableUI}
\title{Phenotype table UI module}
\usage{
phenotypeTableUI(id)
}
\arguments{
\item{id}{namespace identifier}
}
\value{
fluidRow holding the ui code
}
\description{
Phenotype table UI module
}
\examples{
phenotypeTableUI("phenotype_id")

}
\author{
Janina Reeder
}
