% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{readData}
\alias{readData}
\title{Reads in data}
\usage{
readData(filepath, type = "RDS")
}
\arguments{
\item{filepath}{Relative or absolute file path of data object.}

\item{type}{The type of file to be read; default is "RDS", other options are
"RDATA", "BIOM", "TAB", "CSV"}
}
\value{
An MRexperiment object.
}
\description{
This function reads in an MRexperiment object saved as an RDS file, a Biom
file, or a tab - delimited count matrix with features as rows
and samples as columns.
}
