% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intra_module.R
\name{relAbundanceUI}
\alias{relAbundanceUI}
\title{Relative abundance plot module - UI}
\usage{
relAbundanceUI(id)
}
\arguments{
\item{id}{namespace identifier}
}
\value{
box containing the ui code
}
\description{
Relative abundance plot module - UI
}
\author{
Janina Reeder
}
