% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_module.R
\name{reportList}
\alias{reportList}
\title{Report tab module server}
\usage{
reportList(
  input,
  output,
  session,
  dataSource,
  preprocessRep,
  qcRep,
  analysisRep,
  aggIndex,
  reset
)
}
\arguments{
\item{input}{module input}

\item{output}{module output}

\item{session}{app session}

\item{dataSource}{R code to obtain data for rendering}

\item{preprocessRep}{R code containing preprocessing steps of data}

\item{qcRep}{R Code to generate QC plots}

\item{analysisRep}{R Code to generate all analyses saved to reports}

\item{aggIndex}{boolean value representing aggregation steps in analysisRep}

\item{reset}{boolean reactive which resets the module if TRUE}
}
\description{
Report tab module server
}
\author{
Janina Reeder
}
