% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_module.R
\name{reportListUI}
\alias{reportListUI}
\title{report tab ui}
\usage{
reportListUI(id)
}
\arguments{
\item{id}{namespace identifier}
}
\value{
fluidRow holding ui elements
}
\description{
report tab ui
}
\examples{
reportListUI("reportlist_id")

}
\author{
Janina Reeder
}
