% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportRow_module.R
\name{reportRow}
\alias{reportRow}
\title{Report Row}
\usage{
reportRow(input, output, session, type, content)
}
\arguments{
\item{input}{module input}

\item{output}{module output}

\item{session}{app session}

\item{type}{boolean indicating whether checkbox should be included}

\item{content}{R code to show}
}
\value{
reactive boolean indicating whether row is selected
}
\description{
Report Row
}
\author{
Janina Reeder
}
