% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportRow_module.R
\name{reportRowUI}
\alias{reportRowUI}
\title{Report row module consisting of a checkbox, image and description/R code area}
\usage{
reportRowUI(id, type)
}
\arguments{
\item{id}{namespace identifier}

\item{type}{boolean indicating if a selector checkbox should be added}
}
\value{
div holding the UI code
}
\description{
Report row module consisting of a checkbox, image and description/R code area
}
\author{
Janina Reeder
}
