% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{runMicrobiomeExplorer}
\alias{runMicrobiomeExplorer}
\title{Main function to start the Microbiome Explorer Shiny app via a command line
call}
\usage{
runMicrobiomeExplorer()
}
\value{
the shiny application
}
\description{
Main function to start the Microbiome Explorer Shiny app via a command line
call
}
