% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataManipulationFunctions.R
\name{updateSampleInfo}
\alias{updateSampleInfo}
\title{Update Sample Information in MicrobiomeData Object}
\usage{
updateSampleInfo(object, new_sample_info)
}
\arguments{
\item{object}{MicrobiomeData object to update sample information.}

\item{new_sample_info}{New sample information (data frame).}
}
\value{
Updated MicrobiomeData object.
}
\description{
Updates the sample information in a MicrobiomeData object.
}
