% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataValidationFunctions.R
\name{validateSampleInfo}
\alias{validateSampleInfo}
\title{Validate Sample Information}
\usage{
validateSampleInfo(sampleInfo, data)
}
\arguments{
\item{sampleInfo}{A data frame containing sample information, expected to have rows as samples and columns as metadata fields.}

\item{data}{A matrix or data frame of the corresponding data (16S rRNA, metagenomic, or metatranscriptomic).}
}
\value{
A logical value indicating whether the sample information is valid.
}
\description{
Verifies if the sample information is correctly formatted and consistent with the provided dataset.
Ensures that sample identifiers are unique and match those in the dataset.
}
