% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRareFaction.R
\name{AdvancedRarefactionPlot}
\alias{AdvancedRarefactionPlot}
\title{Plot Rarefaction Curves for Microbiome Data}
\usage{
AdvancedRarefactionPlot(
  data,
  step = 1,
  sample = NULL,
  xlab = "Sample Size",
  ylab = "Species Richness",
  col = NULL,
  lty = 1,
  statistical_analysis = FALSE,
  save_plot = FALSE,
  file_name = "rarefaction_plot",
  file_type = "pdf",
  ...
)
}
\arguments{
\item{data}{A matrix or dataframe with rows as samples and columns as taxa.}

\item{step}{The step size for rarefaction curve calculation.}

\item{sample}{The number of individuals to be sampled for each step (if not specified, it uses the minimum sample size).}

\item{xlab}{The label for the x-axis.}

\item{ylab}{The label for the y-axis.}

\item{col}{A vector of colors for the curves, recycled if necessary.}

\item{lty}{The line type for the curves, recycled if necessary.}

\item{save_plot}{Logical, if TRUE, the plot is saved to the specified file.}

\item{file_name}{The name of the file to save the plot to.}

\item{file_type}{The type of the file to save the plot to (e.g., "png", "pdf").}

\item{...}{Additional arguments to pass to the plot function.}
}
\value{
A plot object showing rarefaction curves for each sample.
}
\description{
This function generates rarefaction curves for microbiome count data, useful for comparing species richness among samples.
}
\details{
\itemize{
  \item \strong{Customizable Sampling}: Control over the number of samples and step size for curve generation.
  \item \strong{Plot Customization}: Allows customization of plot aesthetics such as color, line type, and legend.
  \item \strong{Export Option}: Option to save the plot to a file in various formats.
}
}
\examples{
data(microbiome_example) # Load the example dataset
AdvancedRarefactionPlot(microbiome_example, sample = 100,
col = c("red", "blue", "green"), save_plot = TRUE,
file_name = "advanced_rarefaction_curves", file_type = "pdf")

}
\references{
Gotelli, N.J. & Colwell, R.K. (2001). Quantifying biodiversity: procedures and pitfalls in the measurement and comparison of species richness. Ecology Letters, 4(4), 379-391.
}
