% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiomeDataClass.R
\name{getData}
\alias{getData}
\alias{getData,MicrobiomeData,character-method}
\title{Retrieve Data from a MicrobiomeData Object}
\usage{
getData(object, dataType)

\S4method{getData}{MicrobiomeData,character}(object, dataType)
}
\arguments{
\item{object}{A MicrobiomeData object.}

\item{dataType}{A character string specifying the type of data to retrieve.
Valid types are 'rRNA16S', 'Metagenomic', 'Metatranscriptomic', 'SampleInfo'.}
}
\value{
The requested data stored in the MicrobiomeData object.
}
\description{
Extracts specified data from a MicrobiomeData object. This can be used to access
different types of stored data within the object.
}
