% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{microbiome_example}
\alias{microbiome_example}
\title{Microbiome Example Dataset}
\format{
A data frame with 30 rows and 10 variables (taxa).
}
\source{
Generated using createMicrobiomeExample function.
}
\usage{
microbiome_example
}
\description{
A dataset representing hypothetical microbiome count data, useful for examples in functions like AdvancedRarefactionPlot.
}
\keyword{datasets}
