% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiomeDataClass.R
\docType{class}
\name{MicrobiomeData-class}
\alias{MicrobiomeData-class}
\title{MicrobiomeData S4 Class}
\description{
An S4 class for storing and managing microbiome data. It encompasses various
types of microbiome-related data including 16S rRNA gene sequencing data,
metagenomic, and metatranscriptomic data, along with associated sample information.
}
\section{Slots}{

\describe{
\item{\code{rRNA16S}}{16S rRNA gene sequencing data stored as a matrix or dataframe.}

\item{\code{Metagenomic}}{Metagenomic data stored as a matrix or dataframe.}

\item{\code{Metatranscriptomic}}{Metatranscriptomic data stored as a matrix or dataframe.}

\item{\code{SampleInfo}}{A dataframe containing information about the samples.}
}}

