% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiomeDataClass.R
\docType{class}
\name{MicrobiomeData-class}
\alias{MicrobiomeData-class}
\title{MicrobiomeData S4 Class}
\description{
A class for storing and managing microbiome data including 16S rRNA,
metagenomic, and metatranscriptomic data along with sample information.
}
\section{Slots}{

\describe{
\item{\code{rRNA16S}}{16S rRNA gene sequencing data}

\item{\code{Metagenomic}}{Metagenomic data}

\item{\code{Metatranscriptomic}}{Metatranscriptomic data}

\item{\code{SampleInfo}}{Information about the samples}
}}

