% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMicrobiomeExplorer.R
\name{runMicrobiomeExplorerApp}
\alias{runMicrobiomeExplorerApp}
\title{Launch Shiny App for MicrobiomeExplorer}
\usage{
runMicrobiomeExplorerApp()
}
\value{
This function does not return a value, but it opens a Shiny application
in the user's default web browser.
}
\description{
This function initiates a Shiny application included with the MicrobiomeExplorer
package. The app provides an interactive user interface for visualizing and
analyzing microbiome data. It supports functionalities such as generating
various plots, performing statistical analyses, and exploring microbiome datasets.
}
\details{
The application code is located in the \code{./inst/shiny-scripts} directory
within the MicrobiomeExplorer package.
}
\examples{
\dontrun{
  # Launch the Shiny app from the MicrobiomeExplorer package
  MicrobiomeExplorer::runMicrobiomeExplorerApp()
}

}
\references{
Grolemund, G. (2015). Learn Shiny - Video Tutorials.
\href{https://shiny.rstudio.com/tutorial/}{Link}
}
