% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiomeDataClass.R
\name{updateData}
\alias{updateData}
\alias{updateData,MicrobiomeData,character-method}
\title{Update Data in a MicrobiomeData Object}
\usage{
updateData(object, dataType, newData)

\S4method{updateData}{MicrobiomeData,character}(object, dataType, newData)
}
\arguments{
\item{object}{A `MicrobiomeData` object to be updated.G}

\item{dataType}{The type of data to update. Valid types include 'rRNA16S', 'Metagenomic',
'Metatranscriptomic', and 'SampleInfo'.}

\item{newData}{The new data to be inserted into the object.}
}
\value{
The updated `MicrobiomeData` object.
}
\description{
Updates specific data within a MicrobiomeData object. This function allows modification
of the existing data slots in the MicrobiomeData object.
}
